Explicitly defined the puppet vardir config value to avoid a bug
According to the puppet documentation, this parameter should be
populated with a default value of /var/lib/puppet (when running as
root), but it fails to do so. For now, I've hard-coded this default
which should be good for the current needs.

The direct symptom prior to the fix was the following:
    Error: Could not intialize global default settings: Error converting
    value for param 'factpath': Could not find value for $vardir

We are still trying to understand what causes the issue.
The following is a related bug found in the puppet tests
https://github.com/rodjek/rspec-puppet/issues/54